requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pyyaml>=6.0
httpx>=0.25.0
pyttsx3>=2.90
vosk-tts>=0.4.0
numpy>=1.24.0
soundfile>=0.13.0
torch>=2.1.0
transformers>=4.35.0
phonemizer>=3.2.1
einops>=0.7.0
pydub>=0.25.1
psutil>=5.9.0
python-multipart>=0.0.9
gunicorn>=21.2.0
uvicorn>=0.27.0
aiohttp>=3.9.0
sentencepiece>=0.1.99
typing-extensions>=4.9.0
kokoro>=0.1.0
huggingface-hub>=0.19.0
scipy>=1.11.0
onnxruntime>=1.16.0
pydantic>=2.5.0
ffmpeg-python>=0.2.0
misaki>=0.1.0
loguru>=0.7.0
